# Optimization For Array Overload of std.conv.parse