@JackStouffer 2017-01-26
> Your comment is just a hunk of code.

This code is the performance benchmark.

> Please explain what you did.

Replaced the dynamic array with an `Appender`.

>this will pissmize performance in some cases.

Please give an example.